requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
pyyaml>=6.0
python-frontmatter>=1.0.0
requests>=2.28.0
pillow>=9.0.0
beautifulsoup4>=4.11.0